+++ testing/serializer.py
@@ -1,6 +1,6 @@
 import pytest
 
-from pyleus.compat import mock
+from pyleus.testing import mock
 from pyleus.storm.serializers.serializer import Serializer
