+++ tests/configuration_test.py
@@ -4,7 +4,7 @@ import pytest
 
 from pyleus import configuration
 from pyleus import exception
-from pyleus.compat import mock
+from pyleus.testing import mock
 
 
 class TestConfiguration(object):
