+++ tests/json_fields_bolt_test.py
@@ -1,8 +1,7 @@
 import pytest
 
 from pyleus.json_fields_bolt import JSONFieldsBolt
-from pyleus.testing import ComponentTestCase
-from pyleus.compat import mock
+from pyleus.testing import ComponentTestCase, mock
 
 
 class TestJSONFieldsBolt(ComponentTestCase):
